fix: cleanups b/c of merging main branch

+++ crates/ibc-testkit/src/testapp/ibc/core/types.rs
@@ -945,7 +945,7 @@ mod tests {
     fn test_router() {
         #[derive(Debug, Default)]
         struct FooModule {
-            counter: usize,
+            counter: u64,
         }
 
         impl Module for FooModule {
